backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
python-multipart
google-generativeai
gtts
python-dotenv
pillow
httpx
requests
